Import into IntelliJ as Gradle build.
