+++ src/main/java/mr/go/sgfilter/DataFilter.java
@@ -3,11 +3,11 @@ package mr.go.sgfilter;
 /**
  * This interface represents types which are able to filter data, for example:
  * eliminate redundant points.
- * 
+ *
  * @author Marcin Rzeźnicki
  * @see SGFilter#appendPreprocessor(Preprocessor)
  */
 public interface DataFilter {
 
-	double[] filter(double[] data);
+  double[] filter(double[] data);
 }
